pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Latin-Lemmatizer"
version = "0.1.0"
description = "Custom tool for Latin text lemmatisation using CLTK"
readme = "README.md"
authors = [
  {name = "Mitchell G. Harris", email = "mitchopt@gmail.com"}
]
maintainers = [
  {name = "Mitchell G. Harris", email = "mitchopt@gmail.com"}
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Typing :: Typed",
]
license = {text = "MIT"}
dependencies = [
  "typer",
  "rich",
  "pyyaml>=6.0.3",
  "cltk>=1.5.0",
]
requires-python = ">= 3.12"

[dependency-groups]
dev = [
    { include-group = "lint" },
    { include-group = "test" },
    { include-group = "typecheck" },
]
lint = [
    "ruff",
]
test = [
    "coverage",
    "pytest",
]
typecheck = [
    "ty",
]
docs = [
    "zensical",
    "mkdocstrings-python",
]

[project.urls]
bugs = "https://github.com/mitchopt/Latin-Lemmatizer/issues"
changelog = "https://github.com/mitchopt/Latin-Lemmatizer/releases"
documentation = "https://mitchopt.github.io/Latin-Lemmatizer/"
homepage = "https://github.com/mitchopt/Latin-Lemmatizer"

[project.scripts]
latin_lemmatizer = "latin_lemmatizer.cli:app"

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.coverage.run]
branch = true
parallel = true
source = ["src/", "tests/"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 50
exclude_also = [
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@overload",
    "@typing.overload",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    "raise NotImplementedError",
    "\\.\\.\\.",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
package = true
default-groups = ["dev"]
